+++ CHANGES.rst
@@ -41,6 +41,11 @@ Unreleased
     already loaded. :issue:`295`
 -   Do not raise an error for undefined filters in unexecuted
     if-statements and conditional expressions. :issue:`842`
+-   Add ``is filter`` and ``is test`` tests to test if a name is a
+    registered filter or test. This allows checking if a filter is
+    available in a template before using it. Test functions can be
+    decorated with ``@environmentfunction``, ``@evalcontextfunction``,
+    or ``@contextfunction``. :issue:`842`, :pr:`1248`
 
 
 Version 2.11.3